Added lib/narrow.elv
Use the new experimental `-narrow-read` mode to implement the
location, dir-history and last-cmd modes. This is taken from
https://github.com/elves/elvish/blob/master/eval/narrow.elv,
with some improvements to account for recent syntax changes
(removal of automatic lambda arguments, and changes in the
`splits` syntax). It also includes some hooks of my own to
redraw the prompt after returning from these modes.

This is still work in progress, will try to make it more
usable later.