Make caching optional, improve git performance
New variable `$theme:chain:cache_chain` controls whether the chain must
be generated only on every command (default value of `$true`) or on
every keystroke. Caching is enabled by default because on large repos,
the git commands that are run as part of the prompt can slow down typing
considerably.

Also substituted some of the git commands for faster ones (although the
improvements are not so dramatic, so we leave caching enabled by default
for now).