Improve loading of private.elv
Before, the `private` module was loaded inside the scope of an `if`
statement, which meant that any functions or Elvish variables defined in
there would not be visible (environment variables work fine, since they
get stored in the `E:` namespace anyway). By using a `?(...)` construct,
the module is visible and we also get a status variable about whether it
was loaded correctly.